Slightly better signal handling in with_launch_file

+++ ros2_easy_test/ros2_easy_test/decorators.py
@@ -343,16 +343,18 @@ def with_launch_file(  # noqa: C901
                     except NameError:
                         pass  # Maybe, the launch process was not even started
                     else:
-                        # Signal the child launch process to finish
-                        # This is much like pressing Ctrl+C on the console
-                        ros2_process.send_signal(SIGINT)
-                        try:
-                            # Might raise a TimeoutExpired if it takes too long
-                            return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
-                        except TimeoutExpired:  # pragma: no cover
-                            ros2_process.terminate()
-                            # return_code will be larger than 130
-                            return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
+                        return_code = ros2_process.poll()
+                        if return_code is None:
+                            # Signal the child launch process to finish
+                            # This is much like pressing Ctrl+C on the console
+                            ros2_process.send_signal(SIGINT)
+                            try:
+                                # Might raise a TimeoutExpired if it takes too long
+                                return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
+                            except TimeoutExpired:  # pragma: no cover
+                                ros2_process.terminate()
+                                # return_code will be larger than 130
+                                return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
 
                 # Both SUCCESS (0) or the result code of SIGINT (130) are acceptable
                 return_code_problematic = return_code not in {0, 130}
